backend/requirements.txt
requests
Flask
flask-cors
pyhumps
sqlalchemy
#psycopg2-binary
mysqlclient
